% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{add_extra_days}
\alias{add_extra_days}
\title{Add an extra day at the beginning and the end of datetime sequence
using the last and first day of the data}
\usage{
add_extra_days(dtf)
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}
}
\value{
tibble
}
\description{
Add an extra day at the beginning and the end of datetime sequence
using the last and first day of the data
}
