% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{complete_timeseries_datetime}
\alias{complete_timeseries_datetime}
\title{Fill NA values of a datetime sequence vector}
\usage{
complete_timeseries_datetime(dtf)
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}
}
\value{
tibble
}
\description{
Fill NA values of a datetime sequence vector
}
\examples{
dtf_gaps <- timefully::dtf[c(1, 2, 3, 8, 9, 10), ]
print(dtf_gaps)
complete_timeseries_datetime(dtf_gaps)

}
