% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_time_num_to_period}
\alias{convert_time_num_to_period}
\title{Convert numeric time value to a datetime period (hour-based)}
\usage{
convert_time_num_to_period(time_num)
}
\arguments{
\item{time_num}{Numeric time value (hour-based)}
}
\value{
\code{lubridate::period} vector with hours and minutes corresponding to
the numeric input.
}
\description{
Convert numeric time value to a datetime period (hour-based)
}
\examples{
convert_time_num_to_period(1.5)
convert_time_num_to_period(c(0.25, 2))

}
