% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtf}
\alias{dtf}
\title{Time-series profiles of consumption and production energy data}
\format{
A tibble
\describe{
\item{solar}{Solar PV production (in kW)}
\item{building}{Consumption profile of a building (in kW)}
}
}
\usage{
dtf
}
\description{
This data set is extracted from \code{flextools::energy_profiles}
}
\keyword{internal}
