% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_datetime}
\alias{fill_datetime}
\title{Fill NA values of a datetime sequence vector}
\usage{
fill_datetime(dttm)
}
\arguments{
\item{dttm}{datetime sequence vector}
}
\value{
filled datetime sequence vector
}
\description{
Fill NA values of a datetime sequence vector
}
\examples{
incomplete_seq <- as.POSIXct("2024-01-01 00:00:00", tz = "UTC") + 0:4 * 3600
incomplete_seq[c(2, 3)] <- NA
fill_datetime(incomplete_seq)

}
