% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{format_datetime_axis}
\alias{format_datetime_axis}
\title{Format datetime axis}
\usage{
format_datetime_axis(format_list, locale = "nl-NL")
}
\arguments{
\item{format_list}{(named list) ORDER matters key and type of output as value}

\item{locale}{(str) locale to use in settings for example nl-NL (default)}
}
\value{
js code as a string to change datetime axis

JS function to pass to dyAxis function
}
\description{
Format datetime axis
}
\details{
The \code{format_list} entries map to the JavaScript \code{toLocaleString} options:\tabular{llll}{
   Option \tab Value \tab Description \tab Example (nl-NL) \cr
   year \tab "numeric" \tab Full year \tab "2023" \cr
   year \tab "2-digit" \tab Last two digits of the year \tab "23" \cr
   month \tab "numeric" \tab Month as a number \tab "1" \cr
   month \tab "2-digit" \tab Month as a two-digit number \tab "01" \cr
   month \tab "long" \tab Full month name \tab "januari" \cr
   month \tab "short" \tab Abbreviated month name \tab "jan." \cr
   day \tab "numeric" \tab Day of the month \tab "5" \cr
   day \tab "2-digit" \tab Day with leading zero \tab "05" \cr
   weekday \tab "long" \tab Full weekday name \tab "vrijdag" \cr
   weekday \tab "short" \tab Abbreviated weekday name \tab "vr." \cr
   weekday \tab "narrow" \tab Single-letter weekday \tab "V" \cr
   hour \tab "numeric" \tab Hour (24-hour format by default) \tab "14" \cr
   hour \tab "2-digit" \tab Two-digit hour \tab "14" \cr
   minute \tab "numeric" \tab Minute \tab "30" \cr
   minute \tab "2-digit" \tab Two-digit minute \tab "30" \cr
   second \tab "numeric" \tab Second \tab "15" \cr
   second \tab "2-digit" \tab Two-digit second \tab "15" \cr
}
}
\examples{
\dontrun{

 dyplot <- plot_ts(dtf)
 xaxis_format <- format_date_axis(
   list(day = "2-digit", month = "short"),
   locale = "nl_NL"
 )
 dyplot |>
   dyAxis("x", axisLabelFormatter = JS(xaxis_format))

}
}
