% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{format_time_axis}
\alias{format_time_axis}
\title{Format time axis}
\usage{
format_time_axis(format_list, locale = "nl-NL")
}
\arguments{
\item{format_list}{(named list) ORDER matters key and type of output as value}

\item{locale}{(str) locale to use in settings for example nl-NL (default)}
}
\value{
js code as a string to change datetime axis
}
\description{
Format time axis
}
\keyword{internal}
