% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_datetime_seq}
\alias{get_datetime_seq}
\title{Date time sequence with time zone and resolution}
\usage{
get_datetime_seq(start_date, end_date, tzone, resolution)
}
\arguments{
\item{start_date}{Date, start date of the output datetime sequence}

\item{end_date}{Date, end date of the output datetime sequence (included)}

\item{tzone}{character, desired time-zone of the datetime sequence}

\item{resolution}{integer, interval of minutes between two consecutive datetime values}
}
\value{
vector of datetime values
}
\description{
Date time sequence with time zone and resolution
}
\examples{
get_datetime_seq(
  start_date = as.Date("2024-01-01"),
  end_date = as.Date("2024-01-03"),
  tzone = "UTC",
  resolution = 120
)

}
