% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolution.R
\name{get_timeseries_resolution}
\alias{get_timeseries_resolution}
\title{Return the time resolution of a time series dataframe}
\usage{
get_timeseries_resolution(dtf, units = "mins")
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}

\item{units}{character being one of "auto", "secs", "mins", "hours", "days" and "weeks"}
}
\value{
numeric
}
\description{
Return the time resolution of a time series dataframe
}
\examples{
get_timeseries_resolution(dtf, units = "mins")

}
