% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_timeseries_tzone}
\alias{get_timeseries_tzone}
\title{Get the time zone of a time series dataframe}
\usage{
get_timeseries_tzone(dtf)
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}
}
\value{
character
}
\description{
Get the time zone of a time series dataframe
}
\examples{
get_timeseries_tzone(dtf)

}
