% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{get_week_from_datetime}
\alias{get_week_from_datetime}
\title{Week date from datetime value}
\usage{
get_week_from_datetime(dttm)
}
\arguments{
\item{dttm}{datetime vector}
}
\value{
date vector
}
\description{
Week date from datetime value
}
\examples{
dttm <- as.POSIXct(
  c("2024-01-01 08:00:00", "2024-01-02 09:00:00", "2024-01-08 10:00:00"),
  tz = "UTC"
)
get_week_from_datetime(dttm)

}
