% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{get_week_total}
\alias{get_week_total}
\title{Summarise dataframe with weekly total column values}
\usage{
get_week_total(dtf)
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}
}
\value{
tibble
}
\description{
Converts the numeric columns of a time-series data frame to total values per week (sum).
Note that if the input values are in power units (e.g., kW), the output values will be in energy units (e.g., kWh).
}
\examples{
get_week_total(dtf[1:100, ])

}
