% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{get_yearly_datetime_seq}
\alias{get_yearly_datetime_seq}
\title{Yearly date time sequence with time zone and resolution}
\usage{
get_yearly_datetime_seq(year, tzone, resolution)
}
\arguments{
\item{year}{integer, year of the datetime sequence}

\item{tzone}{character, desired time-zone of the datetime sequence}

\item{resolution}{integer, interval of minutes between two consecutive datetime values}
}
\value{
vector of datetime values
}
\description{
Yearly date time sequence with time zone and resolution
}
\examples{
get_yearly_datetime_seq(
  year = 2024,
  tzone = "UTC",
  resolution = 60
)

}
