% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{has_timeseries_gaps}
\alias{has_timeseries_gaps}
\title{Check whether a time series dataframe has gaps}
\usage{
has_timeseries_gaps(dtf)
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}
}
\value{
logical, \code{TRUE} when there are missing time slots
}
\description{
Note that timefully considers a full datetime sequence when days are
complete.
}
\examples{
has_timeseries_gaps(dtf)

dtf_gaps <- dtf[c(1:3, 7:10), ]
has_timeseries_gaps(dtf_gaps)

}
