% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{pad_timeseries}
\alias{pad_timeseries}
\title{Pad a time series to a complete datetime sequence}
\usage{
pad_timeseries(dtf, verbose = TRUE)
}
\arguments{
\item{dtf}{data.frame or tibble, first column of name \code{datetime} being
of class datetime and rest of columns being numeric}

\item{verbose}{logical, when \code{TRUE} (default) report the number of rows
added and the dates with gaps}
}
\value{
tibble
}
\description{
Expands \code{dtf} onto the complete datetime sequence that spans the same date
range at the same resolution, inserting the missing time slots as rows with
\code{NA} values. Note that timefully considers a full datetime sequence when
days are complete.
}
\examples{
# Sample just some hours
dtf_gaps <- dtf[c(1:3, 7:10), ]

# Note that the full day is provided
pad_timeseries(
   dtf_gaps
)
}
