% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{to_hhmm}
\alias{to_hhmm}
\title{Convert a number of minutes in string format "HH:MM"}
\usage{
to_hhmm(mins)
}
\arguments{
\item{mins}{integer, number of minutes (from 0 to 1439)}
}
\value{
character
}
\description{
Convert a number of minutes in string format "HH:MM"
}
\examples{
to_hhmm(75)

}
